move functions after merge
